matching_recommendation/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
Deprecated==1.2.18
dotenv==0.9.9
fastapi==0.115.11
google-generativeai
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hvac
idna==3.10
limits==4.1
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
requests==2.32.3
aiohttp
aiofiles